requirements.txt
# YASA wants numpy 1.x
numpy==1.24.3
nptyping==1.4.4
pandas
brainflow
mne
yasa
influxdb
paho-mqtt
websockets
google-cloud-storage
edfio
firebase-admin
pytz
xgboost
lightgbm
python-dotenv
joblib
catboost
# Last version on Windows that supported GPU.  Also needs no higher than Python 3.10.
tensorflow==2.10
keras
pyedflib
tqdm
paramiko
shap
pylsl
pulearn
unstructured
pdfminer
pymupdf
pydrive
langchain
langchain_community
papermill
scrapbook
imodels
pymongo
firebase-admin
interpret<0.5.0
dash<2.15.0
SALib<1.5.0
soundfile